Create API Service Layer
Implement an API service layer while preserving the existing API functionality and business logic.
[skip gpt_engineer]
