+++ src/hooks/useZipAutocomplete.ts
@@ -1,28 +1,15 @@
 import { useQuery } from "@tanstack/react-query";
-import { supabase } from "@/integrations/supabase/client";
+import { addressService } from "@/services/api/addressService";
 
 export const useZipAutocomplete = (searchTerm: string) => {
   const { data: suggestions = [], isLoading, error } = useQuery({
     queryKey: ['zipSearch', searchTerm],
-    queryFn: async () => {
-      if (!searchTerm) return [];
-      
-      const { data, error } = await supabase.functions.invoke('address-lookup', {
-        body: { 
-          type: 'zip', 
-          searchTerm,
-          limit: 10
-        }
-      });
-
-      if (error) throw error;
-      return data?.QueryAutoComplete4Result?.AutoCompleteResult || [];
-    },
+    queryFn: () => addressService.lookupZip(searchTerm),
     enabled: searchTerm.length > 0,
-    staleTime: 0, // Mark data as immediately stale
-    gcTime: 0, // Don't cache at all (previously cacheTime)
-    refetchOnMount: true, // Always refetch on mount
-    refetchOnWindowFocus: true, // Refetch when window regains focus
+    staleTime: 0,
+    gcTime: 0,
+    refetchOnMount: true,
+    refetchOnWindowFocus: true,
   });
 
   return { suggestions, isLoading, error };
